Add a max content width
Now on larger screens the width of things doesn't get out of control
ridiculous, which makes it hard to read. The max width is a few inches
larger than what it'd be when printed out on standard paper. I could
make it match exactly, but I think it is nice to make use of some of the
extra space on a screen...Just not so much that it stretches out
ridiculously.